+++ raster/r.univar/r3.univar_main.c
@@ -200,7 +200,7 @@ int main(int argc, char *argv[])
         n_zones = 1;
 
     stats = create_univar_stat_struct(map_type, i);
-    for (i = 0; i < n_zones; i++) {
+    for (i = 0; i < (unsigned int)n_zones; i++) {
         unsigned int j;
 
         for (j = 0; j < stats[i].n_perc; j++) {
